+++ internal/plugintest/working_dir.go
@@ -51,6 +51,10 @@ type WorkingDir struct {
 // working directory. After this method is called, the working directory object
 // is invalid and may no longer be used.
 func (wd *WorkingDir) Close() error {
+	if os.Getenv(EnvTfAccPersistWorkingDir) != "" {
+		return nil
+	}
+
 	return os.RemoveAll(wd.baseDir)
 }
